decode working for all opcodes!
This mostly went to plan in vm-instruction-set.txt and AG, but a few minor
changes came up.  Notably it ended up being much simpler to renumber opcodes,
so this change is reflected in vm-instruction-set.txt.

It still looks plausible that we could save one cycle in the high-numbered
opcode path, and the input wiring is already a rats nest, but this feels like a
good checkpoint.

Notes
- AG always triggered decode through p.Ai, which meant for high numbered
  opcodes a constant 5 was used to step p.A to trigger p.G.  This is
  unnecessary, we can just trigger p.Gi directly via a pulse amplifier.
- Since we transmit high numbered opcodes subtractively, instead of using
  p.G{345}o for 7x/8x/9x, it's easier to use p.G{123}o.  So we reserve opcodes
  50-52 instead of 53-55, and then stepper K decodes 7x, J decodes 8x and H
  decodes 9x.
- For the same reason, it's simpler just to use 4-9 instead of 0-5 as the ones'
  digit of high numbered opcodes.
- Bizzarely, we end up having to swap the units' and tens' digits of the opcode
  field when transmitted subtractively (as compared to the additive order)
  to make the p.Gdi input work correctly.