add easm+macro syntax for pulse amps
We'll probably mostly need to allocate individual channels for program lines,
so only support that for starters.  To avoid having to make up names for all
the random buffers, allow very permissive names in reservations and rely on
a $pulseamp macro to concatenate things.  Possibly could be baked into the
assembler itself as a special sort of line but this way was quite easy.