rig constants to fix I1+1 fetch issue
Fix the issue where the first instruction of a line had to be stored as I1-1 by
adding an extra P01 to EX before FT sends 99-I1.  We can get the P01 by wiring
up the constant transmitter "creatively".  I'm like 90% sure this should work
electrically, but it's a little odd.  As a bonus, this gives us a P01 which
might come in useful for inc, and one more free positive constant.