save one cycle in control loop
Originally we needed a fixup cycle 5 to ensure the top of EX was M99, but we
can do better using a trick from AG's setup.  Since EX doesn't transmit sign on
main, we can simultaneously transmit a constant M00000 and have fill99 grab
that to fill the top of IR.

While going through this, I noticed a theoretical defect in the opgt55 path
which I've documented; it looks like we'll luckily dodge the bullet, but do
have to think harder about how to set up our initial conditions correctly.