add minus1 operand correction to chasm
Since we always increment IR after shifting it, instructions that take an
operand will read operand+1.  Fix this by having the assembler emit operand-1
instead.

If the operand is 0, we'll emit 99 and IR+=1 will carry into the next opcode!
So we also need to carry the -1 through any subsequent 99s on the line.  For
example 40 00/35 needs to be 40 99/34.  Or because nops are 00, e.g. 40
00/00/00/35 needs to be 40 99/99/99/34.  This is theoretically a problem for
nops at the end of the line, since e.g. 40 00/00 would give 40 99/99, and
consuming 40 would flip the sign of IR to empty immediately!  But it's probably
semantically fine to skip nops at the end of a line...

In fact this change also pads lines with 99 instead of 00 which has the effect
of immediately skipping to fetch a new line.

Two digit operands are ok here too.  jmp far 300, logically 74 00 99, becomes
74 99 98; this is what we want because we shift IR to consume 74, incrementing
the operand only once to get 00 99 (9900 in IR), then splat that into PC - if
we consumed words from IR one at a time we'd need more complex correction.  And
for jmp far / jsr the rest of the IR doesn't matter.