add JSR/RET instructions
This uses a new auxiliary accumulator, shiftpc, to implement JSR in 6 cycles
and RET in 5 cycles.  To keep JMP xxxx at 5 cycles, continue to use IR directly
for it, even though we could probably consolidate more programs if we used
shiftpc for JMP too.

I tried building JSR on IR since it has enough inputs, but then RET would still
have needed an auxiliary accumulator with separate sequences.  And alas the
clever five cycle scheme for JSR that relied on clearing IR by incrementing
digit 5 ran into trouble with the sign of IR misfiring p-nofetch; so seemed
better to reuse more of RET anyway.

This timing relies on the fact that we can overlap a send of EX.A on main to
restore shiftpc with p-fetchline, since the signs from ftdisc1/2/3 aren't on
main.  Maybe other jmp-taken could benefit?