chasm padding changes
Use 99s for padding everywhere.  Always pad the remainder of the current row
after JSR/JMP/RET because the rest of the instructions are unreachable.

Allow labels to implicitly force row alignment.  Previously, chasm required
an ".align" directive for labels to make it super obvious that they point to
the start of a row, but that syntax is ugly, and we're unlikely to forget.