automate accum dummies
This is mostly transparent to chessvm.easm and is scoped to the $dummy macro.
A "defer" directive defers assembling lines until a later "insert-deferred"
directive, and that's used to move dummies after other accumulator programs in
the output.  A new "allocate-dummy" directive allocates a named register and
transceiver on an allowed set of accumulators, and since this happens after
other programs it won't clash.