add limited conditionals to easm, profile pas+ts
Add 'if', 'endif', and 'define' directives to easm, and use them to optionalize
sections of chessvm.easm.  Use this to profile resource usage by disabling
sections of code; the profile is at the top of the file.  We can also use this
to continue development by turning off some stuff while we figure out how to
get other stuff to fit.