rewrite jmp far/jsr/ret
Saves 4 pas, 4 ts, and frees up one ex input in exchange for a one cycle slower
jmp far.  There is actually a 5 cycle option for all 3 ops in the new shared
fargo sequence, but it would probably use more resources again, so I left a
todo to investigate it more.