renumber opcodes for mov
mov seems ripe for refactoring, but however we do it, we're going to end up
with lots of opcodes that want to share sequences which takes a lot of
pulseamps.  Renumbering opcodes so that the shared sequence movs are all on
the same steppers lets us wire the tens digits to dummies and trade 2 ts to get
back 8 pas.

Presently this leaves 4 unused instruction slots which trigger some mov
programs.  Even if we don't fit more movs into them, we can dummy around the
delays for some longer ops.

Next step is to take a closer look at the 3-cycle mov sequence option and the
slower 6 cycle sequence to see where they land on transceivers.

This lets us bring back clr.