add 6-cycle MOV that shares with SWAP
This uses 7 fewer pas and 4 fewer ts net, compared to 3-cycle MOVs with a
separate SWAP.  But it does fix an issue with memcycle sharing in the current
swapBCDE if we're going to extend memcycle, and also there might be room for
further optimization.

Both paths should still work just fine by switching around "define"s.