vmtests for loadacc/storeacc, if/else, if FAST_MOV
Add if/else to easm and use it to group the enables corresponding to the two
different MOV sequence choices in consideration.  Also support nesting if so we
can condition some nested code that's needed only for one of the two paths
(same resource use though).

add a vmtest for loadacc/storeacc/swapall - currently passes with FAST_MOV
disabled but not with FAST_MOV enabled.