add read instruction
Just to get some basic I/O locked in, will almost certainly change.

Remove the required "AB" argument for print/read, and have read affect LS -
it's less stuff to save/restore, and it sorta makes sense to want to read then
storeacc A.

It'd be nice if we could read 10 digits to fill LS, but ct wiring makes that
cumbersome.  It'd be nice if read/print were interoperable, such that you could
read back cards you've printed... currently we print 0AABB but read FFGGH, so
it's just annoying.