add ftl A,D
this reuses the discjx sequence so costs 1t/1pa, enable it by default.  doing
so requires treating ftl as a sort of always-taken jump, which isn't terribly
efficient, but is quite cheap and easy to improve later if we have space

it's most convenient to compute the index in A, but this is a problem because
then A has the index and ex has discjx during the lookup.  we could try to add
more inputs to rf to clear A before data returns, but it's simpler just to read
data into D instead.

it'd be saner but cost more to first clear D before reading into it - right now
this adds to D, and so may wrap into C etc.