wire up multiplier dummies+pulseamps
The multiplier has deterministic timing, so if we allow ourselves to pretend
its program interconnect connections are left disconnected (eniacsim now
supports this) we can use it for dummies.  Strangely, we can also use its
facility to pass Cpp to receive ier and icand to get a crude sort of pulseamp
replacement.

I have not tested the hybrid pulseamp+dummy macro.