save 11ts/1pa in mov for +2 cycles
We can time things to use the multiplier product disposal programs to save a
bunch of resources for MOV X,A.  This unfortunately makes it even slower -
pushing all the way to 8 cycles.  As an optimization later we could do this
only for LS moves or something, but given how cramped things are, this seems
too good to leave out.