add chasm syntax for implicit swap
We have to write mov X,A/swap A,Y to load Y, so try out some syntax to make
this more concise and explicit.  If we later have resources and want to support
e.g. mov imm,D again, perhaps it'll be easier to port from mov imm,A<->D.