add flipn (and save 1pa/3 dummies)
We can trade some ft dummies for accum dummies in read which is slow enough
that it doesn't matter; also p-rf-selEA is only used from one place now so
doesn't need to be shared so can save one pulseamp.

flipn is perhaps not the best use of the extra resources but seems like a
useful thing that could save some code space.