correct storeacc sign behavior in chsim
Previously chsim just hacked in ft banks for eniac accumulator synchronization,
but this is architecturally observable through loadacc/swapall, so must
actually be simulated at the vm level instead of just patched in.  Also
storeacc overwrites the LS sign before storing, so it's not possible to flip it
by mistake.