reassign movegen+stack addresses
Move wrook2 to 45 so we can fit all the movegen state we desire in a7, which
doubles as top of stack.  Line things up so that we can read out (fromp,
targetp, from, to, movestate) which is nice for debugging.  Also fixup
fen2deck.py and test scripts with the new addresses.