add pawn to queen promotion
Treat a pawn move to rank 1 or 8 as promotion to queen. Done in move.asm with
no awareness in movegen, mostly because there's room there; this way should
permit promotion either by push or capture.

move detects promotion, records it by adding 90 to movestate, and then changes
the pawn to a queen. undo_move checks movestate and reinstates the pawn. mscore
is updated to reflect the value of the queen. This required a small change to
movegen to no longer use 99 as a sentinel value for movestate.

Minimally tested, seems to work and doesn't seem to break anything horribly.