parameterize "quiesence" depth
introduce a param DQ for the depth beyond which we only evaluate captures
if we set it to 3, we can get the runtime around ~5.5 eniac hours/move
which is within our notional mtbf budget (but play is much worse)
this commit has DQ=4, which yields about 10 hours/move and obviously
stronger play. setting DQ>4 (effectively turning off the pruning) is
competitive with stockfish@1350 but takes 20+ hours/move.